fix(ui): swap and sanitize test files
